requirements.txt
SpeechRecognition
google-generativeai
gtts
pydub